# api/openapi.yaml
openapi: 3.0.3
info:
  title: happyDomain API
  description: Finally a simple interface for domain names.
  version: "0.1.0"
  contact:
    email: contact+api@happydomain.org
    name: happyDomain team
  license:
    name: GNU Affero General Public License v3.0 or later
    url: https://spdx.org/licenses/AGPL-3.0-or-later.html
...
servers:

  - url: http://localhost:8081/api
    description: Development server

  - url: https://app.happydomain.org/api
    description: Main production server
paths:
...
  /api/domains/{domainId}/zone/{zoneId}:

    get:
      operationId: get_zone
      description: Retrieve the specified zone with its services and records.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Retrieve Zone information.
      tags:
        - zones

    patch:
      operationId: update_service
      description: Add or update a Service inside the given Zone.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      requestBody:
        description: Service to update
        required: true
        content:
          application/json:
            schema:
              $ref: './schemas.yaml#/components/schemas/Service'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Add or update a Service.
      tags:
        - zones
...
  /api/domains/{domainId}/zone/{zoneId}/{subdomain}/services:

    post:
      operationId: add_service
      description: Add a Service to the given subdomain of the Zone.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      - description: Part of the subdomain considered for the service (@ for the root
          of the zone ; subdomain is relative to the root, do not include it)
        in: path
        name: subdomain
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Add a Service.
      tags:
        - zones
  /api/domains/{domainId}/zone/{zoneId}/{subdomain}/services/{serviceId}:

    delete:
      operationId: delete_service
      description: Drop the given Service.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      - description: Part of the subdomain considered for the service (@ for the root
          of the zone ; subdomain is relative to the root, do not include it)
        in: path
        name: subdomain
        required: true
        schema:
          type: string
      - description: Service identifier
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Zone'
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Drop the given Service.
      tags:
        - zones

    get:
      operationId: get_service
      description: Retrieve the designated Service.
      parameters:
      - description: Domain identifier
        in: path
        name: domainId
        required: true
        schema:
          type: string
      - description: Zone identifier
        in: path
        name: zoneId
        required: true
        schema:
          type: string
      - description: Part of the subdomain considered for the service (@ for the root
          of the zone ; subdomain is relative to the root, do not include it)
        in: path
        name: subdomain
        required: true
        schema:
          type: string
      - description: Service identifier
        in: path
        name: serviceId
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/Service'
        "401":
          description: Authentication failure
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
        "404":
          description: Domain or Zone not found
          content:
            application/json:
              schema:
                $ref: './schemas.yaml#/components/schemas/ErrorResponse'
      security:
        - ApiKeyAuth: []
      summary: Get the Service.
      tags:
        - zones
...
components:
  securitySchemes:

    ApiKeyAuth:
      description: Description for what is this security definition being used
      in: header
      name: Authorization
      type: apiKey

    BasicAuth:
      type: http
      scheme: basic
...